pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semcomf"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A Semantic Communication Framework"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Communications"
]
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "numpy>=1.20.0",
    "Pillow>=8.0.0",
    "loguru",
]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "flake8"]
examples = ["matplotlib", "hydra-core", "pytorch-lightning"]

[project.urls]
"Homepage" = "https://github.com/mokizzz/semcomf"
"Bug Tracker" = "https://github.com/mokizzz/semcomf/issues"